+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Next version
 
 - Refactor: Split out `waku_types` types into right place; create utils folder.
+- Refactor: Replace sequence of ContentTopics in ContentFilter with a single ContentTopic.
 - Docs: Add information on how to query Status test fleet for node addresses; how to view logs and how to update submodules.
 - PubSub topic `subscribe` and `unsubscribe` no longer returns a future (removed `async` designation)
 - Added a peer manager for `relay`, `filter`, `store` and `swap` peers.